Moved auto module assignments to a better place
